+++ src/DevilDaggersInfo.Tools/Ui/Main/MainWindow.cs
@@ -37,7 +37,10 @@ public static class MainWindow
 			ImGui.Text(_version);
 			ImGui.Text("Developed by Noah Stolk");
 
-			ImGui.SetCursorPos(new(windowSize.X - 208, 8));
+			ImGui.SetCursorPos(new(windowSize.X - 260, 8));
+			AppButton(Root.InternalResources.DownloadTexture, "Updates", UiRenderer.ShowUpdate);
+
+			ImGui.SameLine();
 			AppButton(Root.InternalResources.ConfigurationTexture, "Configuration", static () => UiRenderer.Layout = LayoutType.Config);
 
 			ImGui.SameLine();