+++ src/DevilDaggersInfo.Tools/Ui/UpdateWindow.cs
@@ -1,4 +1,5 @@
 using DevilDaggersInfo.Tools.Utils;
+using DevilDaggersInfo.Web.ApiSpec.Tools;
 using ImGuiNET;
 using System.Numerics;
 
@@ -6,8 +7,6 @@ namespace DevilDaggersInfo.Tools.Ui;
 
 public static class UpdateWindow
 {
-	public static Version? AvailableUpdateVersion { get; set; }
-
 	public static void Render(ref bool show)
 	{
 		if (!show)
@@ -17,34 +16,55 @@ public static class UpdateWindow
 		Vector2 windowSize = new(384, 384);
 		ImGui.SetNextWindowPos(center, ImGuiCond.Appearing, new(0.5f, 0.5f));
 		ImGui.SetNextWindowSize(windowSize);
-		if (ImGui.Begin("Update available", ref show, ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoResize))
+		if (ImGui.Begin("Updates", ref show, ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoResize))
 		{
 			ImGui.PushTextWrapPos(windowSize.X - 16);
 
-			// TODO: Refactor this. AvailableUpdateVersion should never be null here.
-			if (AvailableUpdateVersion == null)
-			{
-				ImGui.Text("Internal error getting new version number.");
-			}
-			else
-			{
-				ImGui.Text(Inline.Span($"""
-					Version {AvailableUpdateVersion} is available.
+			ImGui.Text(Inline.Span($"""
+                The current version is {AssemblyUtils.EntryAssemblyVersion}.
 
-					The current version is {AssemblyUtils.EntryAssemblyVersion}.
-					"""));
-				ImGui.Spacing();
+                Unfortunately, since anti-virus software does not like the way automatic updates were implemented, the auto-update feature has been removed.
 
-				ImGuiExt.Hyperlink("https://github.com/NoahStolk/ddinfo-tools/releases", "Download from GitHub");
-				ImGui.Spacing();
+                You can download the latest version from the GitHub releases (link below).
 
-				ImGuiExt.Hyperlink("https://github.com/NoahStolk/ddinfo-tools/blob/main/CHANGELOG.md", "View the full changelog");
-				ImGui.Spacing();
-			}
+                1. Go to the the GitHub releases page.
+                2. Under "Assets", download the {GetZipAssetDisplayName()} file.
+                3. Extract the contents.
+                4. Run {GetExecutableDisplayName()}.
+
+                Your user settings will be preserved.
+                """));
+			ImGui.Spacing();
+
+			ImGuiExt.Hyperlink("https://github.com/NoahStolk/ddinfo-tools/releases", "Download from GitHub");
+			ImGui.Spacing();
+
+			ImGuiExt.Hyperlink("https://github.com/NoahStolk/ddinfo-tools/blob/main/CHANGELOG.md", "View the full changelog");
+			ImGui.Spacing();
 
 			ImGui.PopTextWrapPos();
 		}
 
 		ImGui.End(); // End Update available
+	}
+
+	private static string GetZipAssetDisplayName()
+	{
+		return Root.PlatformSpecificValues.AppOperatingSystem switch
+		{
+			AppOperatingSystem.Windows => "ddinfo-tools-win-x64.zip",
+			AppOperatingSystem.Linux => "ddinfo-tools-linux-x64.zip",
+			_ => "zip",
+		};
+	}
+
+	private static string GetExecutableDisplayName()
+	{
+		return Root.PlatformSpecificValues.AppOperatingSystem switch
+		{
+			AppOperatingSystem.Windows => "ddinfo-tools.exe",
+			AppOperatingSystem.Linux => "ddinfo-tools",
+			_ => "the executable",
+		};
 	}
 }