+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## [0.11.0.0] - 2024-01-17
 
 ### Practice